Check ZONEMD records when loading or transferring a zone
If ZONEMD is present in a zone and check-zonemd is active,
check the hash of the zone contents against the ZONEMD record.

When loading primary zones only, an expired signature is permitted
for the ZONEMD RRset.  This will allow the zone to be re-signed
by named.

When loading or transferring a mirror zone, the zone will now
be DNSSEC validated only if there was no signed ZONEMD to check.
(Note that ZONEMD is more protective than DNSSEC validation,
because it covers glue as well as in-zone data.)