Dynamically allocate the names in the recparam from dns_message pool
The ns_query_t structure has a mechanism to check whether we are not in
a recursion loop.  Previously, the full qname and qdomain was stored in
the structure as dns_fixedname_t which blew up the size of the
ns_query_t by 1kB and 16 cachelines.  Replace the two fixed names with
dynamically allocated messages from the dns_message name pool.