Replace the two dns_fixedname_t in ns_query_t with a single hash
The ns_query_t structure has a mechanism to check whether we are not in
a recursion loop.  Previously, the full qname and qdomain was stored in
the structure as dns_fixedname_t which blew up the size of the
ns_query_t by >= 512 bytes.  Replace the two fixed names and the qtype
with a single 64-bit SipHash of (qtype, qname, qdomain), reducing the
recparam from ~1 kB to 8 bytes.