Reject RRsets that are too big to be returned
The maximum DNS message size is 65535 bytes.  Taking off the DNS
message header, a question section and EDNS with a DNS COOKIE there
are 65225 bytes available with a maximum sized DNS query name.

The committed code does not account for the use of compression
pointers in types which support compression pointers.  In theory
larger rdataslabs could fit in the available 65225 bytes.