Fix the glue table in the QP and RBT zone databases
When adding glue to the header, we add header to the wait-free stack to
be cleaned up later which sets wfc_node->next to non-NULL value.  When
the actual cleaning happens we would only cleanup the .glue_list, but
the wfc_node was not reinitialized, which would lead to the assertion
failure when we would call the addglue() on the same header that had
glue in a version that was cleaned.

Refactor the glue_list to use an extra RCU-enabled gluenode structure
that is used to keep the relationship between the header and the glue,
so that we can keep just a single pointer linked in the dns_slabheader_t
structure that we can atomically swap and then clean without affecting
the header itself.