Rewrite the database API to not expose the node-based design
The current zone and cache database design is node-centric, where each
node corresponds to a DNS name.  Rewrite the dns_db API in a way that
tries to hide this design as much as possible and uses the name-type to
manipulate the rdatasets (findrdataset, addrdataset, subtractrdataset
and removerdataset), and just uses the name to index everything
else (like the iterator acquire with allrdatasets).