QUIC/TLS interaction interface and utilities
This commit provides an interface for interacting with a QUIC
integration TLS functionality. The interface is designed in such a way
that allows multiple QUIC integration APIs co-exist in one
codebase. It is modelled after the QUIC integration API found in
BoringSSL/LibreSSL/QuicTLS and thus makes it easy to make an
implementation of the interface for these libraries.

Two implementations are planned:

1. A "native" one that uses the QUIC integration API provided by
crypto libraries (LibreSSL/QuicTLS).
2. A portable compatibility layer.

An API that allows switching between the two at runtime is meant to
aid in unit and interoperability testing.