don't show expiry time when dumping expired records
previously, when dumping cache contents, expired records
awaiting cleanup were displayed with their time of expiration.
this was done internally by overloading the rdataset->ttl
field to store an absolute time value, instead of a
time-to-live.

this method turns out to be fragile. under some circumstances,
it was possible for an rdataset to be returned in a query
response when the ttl field was set to an absolute time,
which would then be sent to the client as an implausibly
long TTL.

this bug could be fixed in other ways, but the practical value
of dumping the expiry date is fairly low anyway, so we're fixing
it by removing the feature.