Fix test for EDE 1
One of the test cases was invalid as it disabled an algorithm for a name
that is below the zone apex. As this use case is flawed, the behavior
is undocumented, the test should ensure that the disabled algorithm
is at the zone origin.

This commit fixes that. In addition, it adds checks to ensure the
response status is NOERROR when the answer is treated as insecure.

Furthermore, when signing the zones for ds-unsupported and
algorithm-disabled, there is no need to create a KEY record for
'cnameandkey' and 'dnameandkey', this was an accidental copy paste
error. This seems to be also true for the reference to ns8.