Add a test for excluding EDE 1 and 2
Add a test where there exist DS with unsupported digests but
also digests that the resolver does know about. The EDE 2 (Unsupported
DS Digest Type) should not be added.

The same test covers the case where there exist a DNSKEY with
unsupported algorithm, but also a DNSKEY with a supported algorithm.
Similarly, the EDE 1 (Unsupported DNSKEY Algorithm) should not be
added.