Split the ZONEMD checks into two parts
The dns_zone_checkzonemd() function has been split into two parts:
dns_zone_checkzonemd() now checks the validity of the ZONEMD
hash value (if found in the zone) against the zone's database
contents. It can be run incrementally on each part of an IXFR,
and in an offloaded worker thread.

dns_zone_validatezonemd() checks that either a) ZONEMD exists and is
validly signed, b) ZONEMD is validly proven not to exist, or c) there
is no secure delegation in the parent.

The reason for the separation is that dns_zone_validatezonemd() may need
data from the cache or other local zones in order to carry out the
DNSSEC validation process, and such lookups require the view to be
frozen. But a view can be thawed when changes are being made to the
zone table by "rndc addzone" or catalog zone updates. It's possible for
a transfer to be in progress when this occurs, because transfers run on
offloaded work threads instead of regular event loops (which would have
been paused while the view was being changed).

So, dns_zone_checkzonemd() can be run while applying the transfer
(in axfr_apply() or ixfr_apply()), and dns_zone_validatezonemd()
can be run afterward (in axfr_apply_done() or ixfr_apply_done()).