Fix spurious missing key files log messages
I suspect this happens because old key is purged by one zone view,
then the other is freaking out about it.

Keys that are unused or being purged should not be taken into account.

The keyring is maintained per zone. So in one zone, a key in the
keyring is being purged. The corresponding key file is removed.

The key maintenance is done for the other zone view. The key in that
keyring is not yet set to purge, but its corresponding key file is
removed. This leads to "some keys are missing"

I think we should not check the purge variable at this point, but the
current time and purge-keys duration. That is what this commit does.