Use automatic cleanup for dst__openssl_fromlabel()
Refactor the OpenSSL 3.0 version of the dst__openssl_fromlabel()
function to use __attribute__((__cleanup__)) for automatic release of
the resources.

In the process, it was discovered that the function actually could leak
memory when OSSL_STORE_INFO_get1_PUBKEY() would first succeed and then
in the later iterations it would fail.