Improve hashing algorithm and order
This commit changes the order in which substrings are inserted in the
hashmap during name compression. Before, substrings were inserted from
shorted to longest, now it is the opposite.
This allows to terminate the insertion process earlier, making name
compression faster.

It also removes the temporary case conversion, as this is incompatible
with the new insertion order.

Together, these changes lead to a 6.5% speedup for the root zone test in
perflab.