Remove 'fetches-per-zone' configuration option
Both fetches-per-server and fetches-per-zone were written as
experimental features to see what could be done to limit the impact of
pseudo-random subdomain attacks.  The 'fetches-per-server' is more
accurately targeted at the real problem, which is a server that can't
respond fast enough to keep up with the traffic.

Remove 'fetches-per-zone' configuration option in favor of
'fetches-per-server' and less complexity in the resolver code.