Prevent the source port depletion on Linux by using socket option
Since 2015, Linux has introduced a new socket option to overcome TCP
limitations: When an application needs to force a source IP on an active
TCP socket it has to use bind(IP, port=x).  As most applications do not
want to deal with already used ports, x is often set to 0, meaning the
kernel is in charge to find an available port.  But kernel does not know
yet if this socket is going to be a listener or be connected. This new
socket option ask the kernel to ignore the 0 port provided by
application in bind(IP, port=0) and only remember the given IP
address. The port will be automatically chosen at connect() time, in a
way that allows sharing a source port as long as the 4-tuples are
unique.

Enable this new option on the outgoing TCP sockets to overcome this TCP
limitation.