Fix dns_qp_lookup() search end
When looking for names, the search ends with either an exact or a
partial match. These terms make sense when looking for just names.

When looking for a type as well there is the possibility to find an
exact matching name but wrong RR type. This means the name exists but
there is no data for the type.

Add the new case to the lookup function, and introduce a new result
code for this scenario.

Also, the partial match case is slightly different when the type part of
the qpkey may differ. If the offset is somewhere in the trailing type
bytes of the found name, we can be sure that the name part is
equal and there is a parent domain.