Remove erefs from qpznode
With the removal of the opportunistic node cleaning, the external
reference mechanism in qpznode becomes superfluous.

It only ensures that a node that both has changes and has been resigned
is cleaned up only once. But since both cleanings would happen from the
same call of closeversion, and calling clean_zone_node twice in a row
with the same least_serial should be idempotent, the same effect can be
obtained by adding resigned nodes to the cleanup list in closeversion.