Move the memory limit accounting from isc_mem to dns_adb and qpcache
The memory limit accounting was done in isc_mem, but used only in
dns_adb and qpcache units.  Move the memory accounting outside of the
isc_mem unit and add ability to have a single accounting across multiple
memory contexts.

In the future, this will allow the cache and adb to have per-loop memory
contexts with a single memory limit and/or share the global memory
limit between the cache and ADB.