Fallback to TCP immediately on UDP response mismatch
When we receive UDP response that is somehow invalid, assume the server
is under attack and fallback to TCP immediately instead of waiting for
the right response to arrive.  Since we are using connected sockets now
the chances that this situation happens under normal circumstances is
close to zero and thus this is safe thing to do and improves the
resilience against cache poisoning attacks.

Additionally, resovler also kept looking for "better" answer over TCP,
this has been also disabled.  Now, if an answer that is invalid in some
regard (we test for an empty QUESTION section) is received, the resolver
fetch is immediately canceled rather than waiting and hoping a correct
answer will be received somehow.