remove dns_db_attachversion()
dns_db_attachversion() was used in only two places: xfrout
and masterdump, to preserve a pointer to the database version
being dumped or transferred across event loop ticks.

in nearly all cases, the version being attached was the the
database's current version, and in the one case (in dnssec-signzone)
where an uncommitted version was being dumped, it's easy to commit
the version before dumping. therefore, we can just use
dns_db_currentversion() for all of these cases, and the
dns_db_attachversion() API call can be removed.