Remove the qpznode .dirty flag
The qpznode cleaning happens now only when we are closing a version on a
"changed" list.  This can be detected in the qpzone_release() because
the least_serial has a non-zero value.  This mechanism makes the .dirty
flag in the qpznode useless as we already know what nodes needs to be
cleaned up and the .dirty flag was just checking recording this for the
second time.

Remove the .dirty flag from the qpznode and only use the least_serial
non-zero value to decide if we need to call clean_zone_node().