Delete empty leaf nodes from the qpzone
Currently, the qpzonedb doesn't delete empty nodes (at all) as the
previous attempts were broken.  Delete empty nodes that has .wild flag
set to false are are not any of the origin nodes.  This still keeps all
the nodes created by `addwildcards()` in the database, but at least, the
DNS-tree leaf nodes are deleted now.