cfg_aclconfctx_t object is part of named_server
`named_g_actconfctx` is a global variable holding the ACL configuration
context alive (in particular, to dynamically load zones). However, this
object is build once per configuration (early) and is used only inside
server.c `apply_configuration` flow. (Two exceptions: the shutdown flow,
still in server.c and plugin check flow, which doesn't need it, so it's
NULL in such case).

Instead of leaving this global publickly exposed, it is now part of the
`named_server_t` object. This enable to clearly see that the new
instance of the ACL context when reconfiguring the server is known only
by the newly built object and not currently used by "production" object;
and will help to move move logic before the exclusive mode is taken.

The other advantage is that the ACL configuration context can now be
built before the exclusive lock as well.