Add support for DELEG (DE) flag in messages
The DE bit is now correctly interpreted when encountered on the wire.
dig can send the bit by specifing "dig +deflag" or "dig +de".

In querylog output, the DE flag is indicated with the letter "G".

When the DE flag is sent in a query, it is echoed back in the
response to the client, similar to DO.