Add support for the ADT flag in DNSKEYs
An ADT flag in a DNSSEC key will be used to indicate to validators
that they must check for the nonexistence of a DELEG record before
accepting an NS delegation.

The flag can now be set with the "-fA" option in dnssec-keygen or
dnssec-keyfromlbel, and it will be set on all keys generated based on
a dnssec-policy, either inside named or via "dnssec-keygen -k".

note: the ADT flag being added by default to keys generated by
dnssec-policy, but not to keys generated using dnssec-keygen
command line paramters, caused several tests to break due to keys
not having the expected flags. these include kasp, migrate2kasp,
multisigner, nsec3, and various rollover-* tests.

this could have been corrected by changing the dnssec-keygen calls to
include the "-fA" flag, but then we wouldn't be correctly testing
rollovers from old keys to new ones.  so instead, test code that checks
key correctness has been modified to expect more possible key flag
values: 256 for an old-style ZSK, 258 for a new one; 257 for an
old-style KSK, 259 for a new one.

an old-style key can be specified by adding the "noadt" keyword to
key properties in python tests, or "noadt" to the set_keyrole arguments
in shell tests.