Authoritative DELEG support
Support for the DELEG type has been added to the load-time zone
integrity checks, DNSSEC code that identifies zone cuts and closest
enclosers, dnssec-signzone, and the qpzone database.

If a DELEG RRset is present in a zone, then a query with DE=1 for
data below its owner name will be answered with a DELEG referral.

If an NS referral is present and DELEG is not, an NS referral is
returned, along with a proof of DELEG nonexistence if the zone is
signed.

If both NS and DELEG are present, NS is used if DE=0, and DELEG if
DE=1.

Several new special cases have been added to the handling of queries
for data *at* a zone cut; the correct response depends on whether the
delegation record is NS, DELEG, or both, and whether the DE flag is
set.  qpzone_find() and qpzone_check_zonecut() have been refactored
and revised to handle all of these cases correctly:

- For query type KEY, NSEC, or DS, or query type DELEG if DE=1, or
  if the GLUEOK flag is set, we treat the node as if it were not a
  zone cut and search for a normal answer.

- For query type DELEG if DE=0, if NS is present, we return an NS
  referral.  If NS is not present, we treat the node as i it were
  not a zone cut and search for a normal answer.

- For any other query type:
  - If DELEG is present and DE=1: return DELEG referral
  - If NS is not present, but DELEG is, and DE=0: return NXRRSET
  - If NS is present, and either DELEG is not present *or* DE=0:
    return NS referral.