Attach/detach the qpcnode to qpcache only once
Instead of attaching and detaching from the qpcache on every time nobody
is using the node, attach to the qpcache only when node is created and
detach from it when the node is destroyed.  This required the QP tree to
be destroyed when dns_db_destroy() is called as that will destroy the
last internal reference to the node and allows the whole qpcache to be
destroyed.