Fix autosign system test
The prepub.example case is testing the signing records when a
prepublication ZSK rollover is performed. For this we expect the
zone to be in a valid DNSSEC state, so set DS publication timing
metadata in the past.

This means that in the logs, one key less will be logged as being
active (because with the new timing metadata, it was activated in
the past).