Remove version.bind, hostname.bind and make id.server static
This follows up the cleanup of the CH class queries.  Remove the BIND 9
specific version.bind and hostname.bind queries and only keep the
id.server that has to be explicitly configured.

Additionally, remove the `gethostname()` call from the hot-path and only
configure the value at the (re-)configuration time.