Exclude TLDs and root from the max-recursion-queries quota
In the beginning, the cache was without form, and void.  If the query
for CDN hosted content is one of the first queries, it might SERVFAIL
because it CDNs often introduce cross-linking between multiple domains
in multiple TLDs and in the end, a single incoming query can result in
over an hundred of outgoing DNS queries.

Relax the initial limits by excluding all queries initiated for domains
with less than 1 or 2 labels (root and TLDs) from the overall limits.