Reduce the number of outgoing queries
The dns_resolver mode of operation is to resolve all the domains and all
the resolver as it iterates the DNS tree to fill up the cache as quickly
as possible.  This commit reduces the number of outgoing queries by
resolving the nameserver addresses via dns_adb_createfind() serially.
This means there will be only one started ADB find and next ADB find
would only start on pending find timeout or by next fctx_try() call.