Reduce the number of outgoing queries
The dns_resolver mode of operation is to resolve all the domains and all
the resolver as it iterates the DNS tree to fill up the cache as quickly
as possible.  This commit reduces the number of outgoing queries by
reducing the number of remote fetches started for the nameserver
addresses resolution via dns_adb_createfind() to a smaller number per
depth of the recursion since the delegation point.