Suppress UDP retry error messages
The UDP transport code would produce an error message when
encountering UV_EGAIN error when sending data.

That error is a recoverable, soft error that signals that the send
request should be reattempted later and is supposed to be handled by
the higher level code logic. Thus, the code suppresses producing a
largely useless error message in this case.