Allow users of AsyncDnsServer to set AA bit for all responses
Previously, all response had to be set as authoritative explicitely
using DnsResponseSend(..., authoritative=True). After using this it
became obvious that this is obnoxious.

Add an optional keyword-only parameter to AsyncDnsServer that sets the
default value of the AA bit on outgoing responses.

Make all the other parameters keyword-only as well.