make cfg_type declarations less verbose
when more than one cfg_type declaration uses the same set of
methods, we now declare a methods structure separately and
reference that.  so, for example, instead of declaring every
tuple type as using cfg_parse_tuple(), cfg_print_tuple(), and
cfg_doc_tuple(), we can now declare it to use cfg_tuple_methods.

several special-purpose functions have been simplified so that
more types can share method sets. for example, instead of having
separate doc_querysource() and doc_serverquerysource() functions,
which only differed in whether to include the keyword "none", a
single function is now used, with a flag indicating whether "none"
is to be printed or not.