internal memory context for parser/config tree
In order to enforce a dedicated memory context for configuration object
tree (nodes and allocated data inside the nodes), add an internal memory
context private to the parser.

It re-uses the existing pattern to initialize/destroy the memory context
respectively after/before the `isc_lib_initialize()` and
`isc_lib_shutdown()` as this maintains state handling the various memory
contextes.