Select a subset of lightweight jobs to run in a dedicated VM
A few jobs in CI take seconds to run, whereas the environment takes more
than a minute to setup. In order to save time and resources, these jobs
are now handled  by a runner in a dedicated VM. They are marked with the
tag "smalljob".

Small jobs are grouped in the "quick-jobs" stage (previously "precheck"),
while longer runnig jobs are moved to "other-jobs" (former "postcheck").
This required reordering of related jobs in .gitlab-ci.yml.