Add type and covers to dns_db_findnode() call + squash findnsec3node
Previously, the logic whether we are looking up nsec3 or normal node
was left up to the caller.  Add type and covers argument to the
dns_db_findnode() and make that logic internal to the function.  This
also allows squashing dns_db_findnode() and dns_db_findnsec3node() into
a single function.

This change will also make future changes to the database easier as it
will allow the implementations to store the type as part of the key when
looking up the node.  That said there is still much to be done as
sometimes, the code expects all the types to reside at the single node,
so more work is needed in this area before this would be possible.