dnssec-signzone: use the zone file's basename as origin
When the zone origin is not specified using the `-o` parameter,
the default behavior is to try to sign using the zone's file name
as the origin. So, for example, `dnssec-signzone -S example.com`
will work, so long as the file name matches the zone name.

This now also works if the zone is in a different directory.
For example, `dnssec-signzone -S zones/example.com` will set
the origin value to `example.com`.