Relax the rules to accept sibling GLUE
The rules for accepting sibling GLUE were hardened due to CVE-40778.

In 2e40705c06831988106335ed77db3cf924d431f6, a code was added to skip
caching of the positive records from the AUTHORITY section if the DNS
message was received over spoofable channel.

Because CVE-40778 relies on spoofing of the DNS message, we can relax
the rules on caching the sibling GLUE and accept the sibling GLUE if it
was received over channel with spoofing protection, e.g. received either
over TCP, or over UDP but protected with DNS cookie or signed with TSIG.