findzonecut: helper function to lookup the cache
Extract the cache lookup implementation from `dns_view_findzonecut()`
into a separate helper function.

Also, when the cache result is not ISC_R_SUCCESS (which is the only
"success" value from the existing code in this case), the return value
is overriden to DNS_R_NXDOMAIN. This enable the caller (in follow-up
commit) to differenciate the case where a zone is found, but for
whatever reason, no delegation is in there, and the case where no zone
is found. Separating those both cases enable to know when it needs to
hits the cache/hints or not.