findzonecut: helper function to lookup zone
Move the logic looking up a zone and its RR into a separate helper
function. The main extra change here is the return codes. This helper
function can return 3 (and only those 3) values:

- ISC_R_SUCCESS: a delegation is found
- DNS_R_NXDOMAIN: no zone matches the qname
- ISC_R_NOTFOUND: a zone is found, but no delegation in the zone (this
  can be either because there is simply no delegation, or for other
  reasons, like, the zone DB is not loaded yet).

This enable the caller to make decisions, typically, attempt a cache
and/or a hints hit or not.