findzonecut: refactor using helper functions
The main function `dns_view_findzonecut()` is rewritten using the
separate helper functions dedicated to hints, cache and zone lookups.

From caller point of view, the API is slightly simplified: the return
value can only be ISC_R_SUCCESS (a delegation is found) or
DNS_R_NXDOMAIN (no delegation is found). Also, if no delegation is
found, this is guarantee that the rdatasets passed as parameter are
disassociated.